ImageServiceDocker/requirements.txt
pillow
tensorflow
keras_applications
boto3